auth-service/requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.1.0
Werkzeug==3.1.3
requests
pymysql
flask-jwt-extended
pytest
SQLAlchemy==2.0.41
marshmallow==4.0.0
Flask-Limiter==3.12
Flask-Mail==0.10.0

cryptography